django/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.13.1
dnspython==2.2.1
eventlet==0.33.1
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
httplib2shim==0.0.3
idna==3.3
kombu==5.2.4
lxml==4.9.1
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==4.21.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==3.0.9
pyTelegramBotAPI==4.6.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
